JSF: floating point adaptions

+++ cob_vision_utils/common/src/VisionUtils.cpp
@@ -645,7 +645,7 @@ unsigned long ipa_Utils::FilterTearOffEdges(cv::Mat& xyzImage, cv::Mat* mask, fl
 				//vDiff.Normalize();
 				dot = (float)vDiff.ddot(vLeft);
 				//dot = vDiff.Dot(vLeft);
-				angle = std::acosf(dot);
+				angle = (float)std::acos(dot);
 				//angle = Wm4::Math<float>::ACos( dot );
 				if (angle > t_upper || angle < t_lower)
 				{
@@ -679,7 +679,7 @@ unsigned long ipa_Utils::FilterTearOffEdges(cv::Mat& xyzImage, cv::Mat* mask, fl
 				//vDiff.Normalize();
 				dot = (float)vDiff.ddot(vLeft);
 				//dot = vDiff.Dot(vLeft);
-				angle = std::acosf(dot);
+				angle = (float)std::acos(dot);
 				//angle = Wm4::Math<float>::ACos( dot );
 				if (angle > t_upper || angle < t_lower)
 				{
@@ -713,7 +713,7 @@ unsigned long ipa_Utils::FilterTearOffEdges(cv::Mat& xyzImage, cv::Mat* mask, fl
 				//vDiff.Normalize();
 				dot = (float)vDiff.ddot(vLeft);
 				//dot = vDiff.Dot(vLeft);
-				angle = std::acosf(dot);
+				angle = (float)std::acos(dot);
 				//angle = Wm4::Math<float>::ACos( dot );
 				if (angle > t_upper || angle < t_lower)
 				{
@@ -746,7 +746,7 @@ unsigned long ipa_Utils::FilterTearOffEdges(cv::Mat& xyzImage, cv::Mat* mask, fl
 				//vDiff.Normalize();
 				dot = (float)vDiff.ddot(vLeft);
 				//dot = vDiff.Dot(vLeft);
-				angle = std::acosf(dot);
+				angle = (float)std::acos(dot);
 				//angle = Wm4::Math<float>::ACos( dot );
 				if (angle > t_upper || angle < t_lower)
 				{
